+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "edu.cascadia.emilio.foncfphoto_points"
+        applicationId "edu.cascadia.mobas.northcreekforest"
         minSdkVersion 19
         targetSdkVersion 28
         versionCode 1
