+++ app/src/androidTest/java/edu/cascadia/mobas/northcreekforest/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
