+++ app/src/main/java/edu/cascadia/mobas/northcreekforest/About.java
@@ -1,4 +1,4 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
 import android.app.Activity;
 import android.os.Bundle;
