+++ app/src/main/java/edu/cascadia/mobas/northcreekforest/MainActivity.java
@@ -1,18 +1,10 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
-import android.app.Activity;
-import android.content.Intent;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.TextView;
 
 
-import edu.cascadia.emilio.foncfphoto_points.dummy.DummyContent;
+import edu.cascadia.mobas.northcreekforest.dummy.DummyContent;
 
 public class MainActivity extends FragmentActivity implements PlanetFragment.OnListFragmentInteractionListener {
