+++ app/src/main/java/edu/cascadia/mobas/northcreekforest/MyPlantsRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -6,22 +6,21 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import edu.cascadia.emilio.foncfphoto_points.PlanetFragment.OnListFragmentInteractionListener;
-import edu.cascadia.emilio.foncfphoto_points.dummy.DummyContent.DummyItem;
+import edu.cascadia.mobas.northcreekforest.dummy.DummyContent.DummyItem;
 
 import java.util.List;
 
 /**
  * {@link RecyclerView.Adapter} that can display a {@link DummyItem} and makes a call to the
- * specified {@link OnListFragmentInteractionListener}.
+ * specified {@link PlanetFragment.OnListFragmentInteractionListener}.
  * TODO: Replace the implementation with code for your data type.
  */
 public class MyPlantsRecyclerViewAdapter extends RecyclerView.Adapter<MyPlantsRecyclerViewAdapter.ViewHolder> {
 
     private final List<DummyItem> mValues;
-    private final OnListFragmentInteractionListener mListener;
+    private final PlanetFragment.OnListFragmentInteractionListener mListener;
 
-    public MyPlantsRecyclerViewAdapter(List<DummyItem> items, OnListFragmentInteractionListener listener) {
+    public MyPlantsRecyclerViewAdapter(List<DummyItem> items, PlanetFragment.OnListFragmentInteractionListener listener) {
         mValues = items;
         mListener = listener;
     }
