+++ app/src/main/java/edu/cascadia/mobas/northcreekforest/PlantInfo.java
@@ -1,6 +1,5 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
-import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
