+++ app/src/main/java/edu/cascadia/mobas/northcreekforest/dummy/DummyContent.java
@@ -1,0 +1,94 @@
+package edu.cascadia.mobas.northcreekforest.dummy;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * Helper class for providing sample content for user interfaces created by
+ * Android template wizards.
+ * <p>
+ * TODO: Replace all uses of this class before publishing your app.
+ */
+public class DummyContent {
+
+    /**
+     * An array of sample (dummy) items.
+     */
+    public static final List<DummyItem> ITEMS = new ArrayList<DummyItem>();
+
+    /**
+     * A map of sample (dummy) items, by ID.
+     */
+    public static final Map<String, DummyItem> ITEM_MAP = new HashMap<String, DummyItem>();
+
+    //private static final int COUNT = 25;
+
+    static {
+        // use this at some point to increment the id
+        // Add some sample items.
+        //for (int i = 1; i <= COUNT; i++) {
+            //addItem(createDummyItem(i));
+        //}
+        addItem(new DummyItem("1", "Douglas Fir", "Pseudotsuga menziesii", "Douglas-fir wood and bark was thought by most of the coastal groups to be an excellent fuel, but it had the reputation of throwing sparks and giving splinters to those handling it. The wood was also used to make items such as spear handles, harpoon shafts, spoons, dip-net poles, harpoon barbs, fire tongs, salmon weirs, caskets and halibut and cod hooks."));
+        addItem(new DummyItem("2", "Black Twinberry", "Lonicera involucrata", "The shiny, black, bitter, twinned berries are not considered edible by most people. They were given names like 'raven's food,' 'crow berry,' and 'monster's food' by northwest coast peoples. There were some taboos against eating them. For example, the Kwakwaka'wakw: believed that eating the berries would cause one to become unable to speak."));
+        addItem(new DummyItem("3", "Red-oiser Dogwood", "Cornus sericea", "Although red-osier-dogwood berries are very bitter and contain a large central seed, they were eaten by interior aboriginal peoples, but not by coastal peoples. The bark was dried and used in smoking mixtures by aboriginal peoples to the East, but not on the Northwest Coast. The branches were sometimes used for salmon spreaders and basket rims."));
+        addItem(new DummyItem("4", "Thimbleberry", "Rubus parviflorus", "Thimbleberries were eaten by all Northwest Coast people. The Nuu-chah-nulth were known to collect shoots in bulk. Young shoots were eaten raw. The berries are coarse and seedy, lending themselves to drying. The Nuu-chah-nulth dried the berries with smoked clams."));
+        addItem(new DummyItem("5", "Sitka Spruce", "Picea sitchensis", "The sharp needles of spruce were believed to give it species paowers for protection against evil thoughts. The Ditidaht and other Nuu-chah-nulth peoples used the boughs in winter dance ceremonies to protect the dancers and to 'scare' spectators. Among the Haida, Tlingit, Tsimshian and other central and northern coastal peoples, the inner bark was eaten fresh or dried into cakes and eaten with berries. The Makah were said to eat the young shoots raw; these would have been an excellent source of vitamin C."));
+        addItem(new DummyItem("6", "Slough Sedge", "Carex obnupta", "Slough sedge is still a popular basket material for the Nuu-chah-nulth and the Makah. The inner leaves are split and flattened before being dried. Fine baskets are made from this sedge, often with the cedar foundations and intricate designs from dyed strands of grass or coloured barks."));
+        addItem(new DummyItem("7", "Clustered Rose", "Rosa pisocarpa", "Branches of all species of wild rose - along with skunk cabbage leaves, fern fronds, pine needles, or salal - were sometimes put in steaming pits, cooking baskets and root-storage pits. Cedar-root cooking baskets used for boiling foods often had rose leaves placed under and over food to flavour it and protect it from burning."));
+        addItem(new DummyItem("8", "Western Redcedar", "Thuja plicata", "Redcedar has been called 'the cornerstone of northwest coast Indian culture' and the large-scale use of its wood and bark delineates the cultural boundary of the northwest coast people within its range. The easily split, rot-resistant wood was used to make important cultural items such as dugout canoes, house planks and posts, totem and mortuary poles, bentwood boxes, baskets, clothing and hats, and a variety of tools and implements such as dishes, arrow shafts, harpoon shafts, spear poles, barbecue sticks, fish spreaders and hangers, dip=net hooks, fish clubs, masks, rattles, benches, cradles, coffins, herring rakes, canoe bailers, ceremonial drum logs, cobs, fishing floats, berry-drying racks, fish weirs, spirit whistles, and paddles."));
+        addItem(new DummyItem("9", "Red Flowering Currant", "Ribes sanguineum", "The berries are edible but insipid. they were eaten by various Coast Salish groups such as the Saanich, Cowichan, Squamish, and Sechelt, but they were not highly regarded. They were eaten fresh but not usually collected for drying. This shrub was introduced to European horticulture by plant-hunter David Douglas. The name sanguineum means 'blood-red' or 'bloody' - rather violent epithets to apply to the beautiful reddish-pink flower, which are harbingers of spring and hummingbirds."));
+        addItem(new DummyItem("10", "Small-fruited Bulrush", "Scirpus microcarpus", "Small-flowered bulrush was apparently used to weave light-duty baskets and as an ornamental trim in hide clothing. The leaves were sometimes laid over and under food in steaming pits"));
+        addItem(new DummyItem("11", "Low Oregon Grape", "Berberis nervosa", "The tart, purple berries of both Oregon-grapes were eaten, but generally not in quantity. Often they were mixed with salal or some other sweeter fruit. Today they are used for jelly, and some people make wine from them. An excellent jelly can be made using two cups Oregon-grape juice, two cups salal juice, five cups sugar and one box pectin crystals."));
+        addItem(new DummyItem("12", "Tall Oregon Grape", "Berberis aquifolium", "The tart, purple berries of both Oregon-grapes were eaten, but generally not in quantity. Often they were mixed with salal or some other sweeter fruit. Today they are used for jelly, and some people make wine from them. An excellent jelly can be made using two cups Oregon-grape juice, two cups salal juice, five cups sugar and one box pectin crystals"));
+        addItem(new DummyItem("13", "Pacific Ninebark", "Physocarpus capitatus", "The Nuu-chah-nulth made children's bows and other small items from the wood, and the Cowichan recently have made knitting needles from it. The Nuxalk considered this shrub highly poisonous and, along with the Coast Salish and Kwakwaka'wakw, used a tea made from a stick with the outer bark peeled off as an emetic or purgative. These three groups also used a medicine from ninebark as a laxative. The Nuxalk used medicinal preparations from ninebark for gonorrhea and scrofulous sores on the neck."));
+        addItem(new DummyItem("14", "Cascara", "Frangula purshiana", "The bark was boiled and the tea (or syrup) was drunk as a strong laxative by the Nuxalk, Coast Salish, Quileute, Nuu-chah-nulth, Kwakwaka'wakw and other groups. Cascara has been scientifically substantiated as an effective laxative. The hydroxymethylanthraquinones it contains cause peristalsis of the large intestine, with little or no effect on the small intestine at low dosages."));
+        addItem(new DummyItem("15", "Mock Orange", "Philadelphus lewisii", "The wood is strong and hard; it never cracks or warps when properly prepared. It is most widely used for making implements among the Interior Salish. On the coast, the Saanich used the wood for bows and arrows, and the Lummi of Washington made combs, netting shuttles and, recently, knitting needles. The Cowlitz also used the wood for combs, and the Skagit made arrow shafts."));
+        addItem(new DummyItem("16", "Pacific Willow", "Salix lasiandra", "The Straits Salish and the Halq'emeylem peeled the bark of [native willow species] in May or June, removed the outer part, split the inner tissue into thin strands, and twisted these into long ropes. This rope was used to make fishing lines and various types of nets, including gill-nets, reef-nets, purse-nets, bag-nets, and duck-nets."));
+        addItem(new DummyItem("17", "Black Cottonwood", "Populus trichocarpa", "The sweet inner bark and cambium tissues were eaten in late spring and early summer immediately after they were harvested so they would not ferment. Hesquiat/Nuu-chah-nulth/Ditidaht: The buds were picked in the spring and boiled in deer fat to make a fragrant salve, which was molded in the bulbous float of bull kelp. Nuxalk: the gum from the buds was used in preparations for baldness, sore throats, whooping cough and tuberculosis...the old, rotten leaves were boiled and used in a bath for body pains, rheumatism and stomach trouble. "));
+        addItem(new DummyItem("18", "Paper Birch", "Betula papyrifera", "Paper birch was used much more commonly by interior groups. The bark, which can be peeled off the tree in large, flexible, waterproof sheets, was as important to the native peoples inland from the Northwest Coast as the western redcedar was to the coastal peoples. Baskets and canoes were the most common items constructed from paper birch. The wood was also used as duel, and the sap and inner bark as emergency food."));
+        addItem(new DummyItem("19", "Grand Fir", "Abies grandis", "Kwakwaka'wakw shamans wove its branches into headdresses and costumes and used the branches for scrubbing individuals in purification rites. The Hesquiat used its branches as incense and decorative clothing for wolf dancers."));
+        addItem(new DummyItem("20", "Red Alder", "Alnus rubra", "Red alder wood is considered to be the best possible fuel for smoking salmon and other types of fish. It is soft and even-grained, and is still used for making feast bowls, masks, rattles, and a variety of other items."));
+        addItem(new DummyItem("21", "Red Elderberry", "Sambucus racemosa", "Red elderberries, though small and seedy, were a highly important food for the peoples of the central and northern coast, although few people sill use them today. They should always be cooked, since the raw berries may cause nausea. They were sometimes boiled to make a sauce or cooked with the stems intact."));
+    }
+
+    private static void addItem(DummyItem item) {
+        ITEMS.add(item);
+        ITEM_MAP.put(item.id, item);
+    }
+
+    //private static DummyItem createDummyItem(int position) {
+        //return new DummyItem(String.valueOf(position), "Item " + position, makeDetails(position));
+    //}
+
+    private static String makeDetails(int position) {
+        StringBuilder builder = new StringBuilder();
+        builder.append("Details about Item: ").append(position);
+        for (int i = 0; i < position; i++) {
+            builder.append("\nMore details information here.");
+        }
+        return builder.toString();
+    }
+
+    // Talk with the team and change this later.
+    public static class DummyItem {
+        public final String id;
+        public final String content;
+        public final String details;
+        public final String info;
+
+        public DummyItem(String id, String content, String details, String info) {
+            this.id = id;
+            this.content = content;
+            this.details = details;
+            this.info = info;
+        }
+
+        @Override
+        public String toString() {
+            return content;
+        }
+    }
+}
