+++ app/src/main/res/layout/activity_about.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".About">
+    tools:context="edu.cascadia.mobas.northcreekforest.About">
 
     <TextView
         android:id="@+id/textView"
