+++ app/src/main/res/layout/fragment_plants_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".PlanetFragment"
+    tools:context="edu.cascadia.mobas.northcreekforest.PlanetFragment"
     tools:listitem="@layout/fragment_plants" />
