+++ app/src/test/java/edu/cascadia/mobas/northcreekforest/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package edu.cascadia.emilio.foncfphoto_points;
+package edu.cascadia.mobas.northcreekforest;
 
 import org.junit.Test;
